# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/service/service_add_http": {
            "post": {
                "description": "添加HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "添加HTTP服务",
                "operationId": "/service/service_add_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "polygon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAddHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_delete": {
            "get": {
                "description": "服务删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务删除",
                "operationId": "/service/service_delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_detail": {
            "get": {
                "description": "服务详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务详情",
                "operationId": "/service/service_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ServiceDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_list": {
            "get": {
                "description": "服务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务列表",
                "operationId": "/service/service_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页个数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServiceListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_stat": {
            "get": {
                "description": "服务统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务统计",
                "operationId": "/service/service_stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DashServiceStatOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_update_http": {
            "post": {
                "description": "修改HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "修改HTTP服务",
                "operationId": "/service/service_update_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "polygon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceUpdateHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dao.AccessControl": {
            "type": "object",
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "open_auth": {
                    "type": "integer"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "white_host_name": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
...
        "dao.GrpcRule": {
            "type": "object",
            "properties": {
                "header_transfor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
        "dao.HttpRule": {
            "type": "object",
            "properties": {
                "header_transfor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "need_https": {
                    "type": "integer"
                },
                "need_strip_uri": {
                    "type": "integer"
                },
                "need_websocket": {
                    "type": "integer"
                },
                "rule": {
                    "type": "string"
                },
                "rule_type": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "url_rewrite": {
                    "type": "string"
                }
            }
        },
        "dao.LoadBalance": {
            "type": "object",
            "properties": {
                "check_interval": {
                    "type": "integer"
                },
                "check_method": {
                    "type": "integer"
                },
                "check_timeout": {
                    "type": "integer"
                },
                "forbid_list": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_list": {
                    "type": "string"
                },
                "round_type": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "upstream_connect_timeout": {
                    "type": "integer"
                },
                "upstream_header_timeout": {
                    "type": "integer"
                },
                "upstream_idle_timeout": {
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "type": "integer"
                },
                "weight_list": {
                    "type": "string"
                }
            }
        },
        "dao.ServiceDetail": {
            "type": "object",
            "properties": {
                "access_control": {
                    "type": "object",
                    "$ref": "#/definitions/dao.AccessControl"
                },
                "grpc_rule": {
                    "type": "object",
                    "$ref": "#/definitions/dao.GrpcRule"
                },
                "http_rule": {
                    "type": "object",
                    "$ref": "#/definitions/dao.HttpRule"
                },
                "info": {
                    "type": "object",
                    "$ref": "#/definitions/dao.ServiceInfo"
                },
                "load_balance": {
                    "type": "object",
                    "$ref": "#/definitions/dao.LoadBalance"
                },
                "tcp_rule": {
                    "type": "object",
                    "$ref": "#/definitions/dao.TcpRule"
                }
            }
        },
        "dao.ServiceInfo": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "load_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.DashServiceStatOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DashServiceStatItemOutput"
                    }
                },
                "legend": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.ServiceAddHTTPInput": {
            "type": "object",
            "required": [
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer"
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string"
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer"
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer"
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer"
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer"
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer"
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string"
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer"
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string"
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer"
                },
                "service_name": {
                    "description": "关键词",
                    "type": "string"
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer"
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer"
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer"
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string"
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string"
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string"
                }
            }
        },
...
        "dto.ServiceListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceListItemOutput"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "dto.ServiceUpdateHTTPInput": {
            "type": "object",
            "required": [
                "id",
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer"
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string"
                },
                "id": {
                    "description": "服务ID",
                    "type": "integer",
                    "example": 62
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string",
                    "example": "127.0.0.1:80"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer"
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer"
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer"
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer"
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer"
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string",
                    "example": "/test_http_service_indb"
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer"
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string",
                    "example": "test_http_service_indb"
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer"
                },
                "service_name": {
                    "description": "服务名",
                    "type": "string",
                    "example": "test_http_service_indb"
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer"
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer"
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer"
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string"
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string",
                    "example": "50"
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string"
                }
            }
        },
...
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}